fits deleted. Entry point added
