+++ setup.cfg
@@ -53,6 +53,10 @@ myraflib =
     py.typed
     sample.fits
 
+[options.entry_points]
+console_scripts =
+    myraf = main:main
+
 [flake8]
 max-line-length = 160
 per-file-ignores = __init__.py:F401
