+++ src/MYRaf.egg-info/SOURCES.txt
@@ -5,6 +5,7 @@ setup.cfg
 src/MYRaf.egg-info/PKG-INFO
 src/MYRaf.egg-info/SOURCES.txt
 src/MYRaf.egg-info/dependency_links.txt
+src/MYRaf.egg-info/entry_points.txt
 src/MYRaf.egg-info/not-zip-safe
 src/MYRaf.egg-info/requires.txt
 src/MYRaf.egg-info/top_level.txt
@@ -14,5 +15,6 @@ src/myraflib/dataarray_model.py
 src/myraflib/error.py
 src/myraflib/fits.py
 src/myraflib/fitsarray.py
+src/myraflib/py.typed
 src/myraflib/sample.fits
 src/myraflib/utils.py
