+++ src/MYRaf.egg-info/entry_points.txt
@@ -1,0 +1,2 @@
+[console_scripts]
+myraf = main:main
